On branch v2 Your branch is up to date with 'origin/v2'.
Changes to be committed:
	modified:   app/Http/Controllers/anggotaController.php
	modified:   app/Http/Controllers/beritaController.php
	modified:   app/Http/Controllers/kepsekController.php
	modified:   app/Http/Controllers/userController.php
	modified:   app/Http/Controllers/wakasekController.php
	modified:   app/Models/anggota.php
	modified:   database/migrations/2023_01_26_012620_create_anggotas_table.php
	deleted:    public/images/dokumentasi/foto-kegiatan/4FN4yOeI65.jpg
	deleted:    public/images/dokumentasi/foto-kegiatan/5xbk7BwgEl.jpg
	deleted:    public/images/dokumentasi/foto-kegiatan/AiTaUVZrJW.jpg
	deleted:    public/images/dokumentasi/foto-kegiatan/wNYCVZaNdk.jpg
	deleted:    public/images/foto-berita/6Ay0Bxu6Ka.jpg
	renamed:    public/tentang-kami/ikhsan.JPG -> public/images/tentang-web/ikhsan.jpg
	new file:   public/images/tentang-web/ikhsan.png
	new file:   public/images/tentang-web/naufal.png
	new file:   public/images/tentang-web/yusuf.png
	modified:   public/webex/css/main.css
	modified:   resources/views/admin/anggota-eskul/anggota.blade.php
	modified:   resources/views/admin/anggota-eskul/editanggota.blade.php
	modified:   resources/views/admin/pendaftaran-eskul/pendaftaran.blade.php
	modified:   resources/views/layout/subnav/pendaftaran-eskul/detail-pendaftaran.blade.php
	modified:   resources/views/layout/subnav/tentangkami/tentang-web-ini.blade.php

+++ app/Http/Controllers/anggotaController.php
@@ -49,7 +49,7 @@ class anggotaController extends Controller
     public function detail_anggota($slug)
     {
 
-        $detail = anggota::where('slug_anggota', $slug)->first();
+        $detail = anggota::where('nis', $slug)->first();
         // dd($detail);
         return view('layout.subnav.detail-anggota', compact('detail'));
     }
@@ -64,30 +64,29 @@ class anggotaController extends Controller
             'jurusan' => 'required',
         ]);
         $data = anggota::create($request->all());
-        $data['slug_anggota'] = $data['nis'];
         $data->save();
         return redirect()->route('anggota')->with('success', ' Data Berhasil Di Tambahkan');
     }
 
     public function editanggota($id)
-    {
+    {   
+        $currentRole = role::where("role", "=", auth()->user()->role)->first();
         $data = anggota::find($id);
         $data_eskul = eskul::all();
-        return view('admin.anggota-eskul.editanggota', compact('data', 'data_eskul'));
+        return view('admin.anggota-eskul.editanggota', compact('data', 'data_eskul','currentRole'));
     }
 
     public function updateanggota(Request $request, $id)
     {
         $request->validate([
-            'nis' => 'required|unique:anggotas',
+            'nis' => 'required',
             'nama_anggota' => 'required',
             'kelas_anggota' => 'required',
             'id_eskul' => 'required',
             'jurusan' => 'required',
-            'slug_anggota' => 'required',
+            
         ]);
         $data = anggota::find($id);
-        $data->slug_anggota = Str::slug($request->get('judul_anggota'));
 
         $data->update($request->all());
 
@@ -97,10 +96,11 @@ class anggotaController extends Controller
 
     public function deleteanggota($id)
     {
-        
         $data = anggota::find($id);
-        
+        if (auth()->user()->id != $data['id_eskul']) {
+            return redirect()->route('anggota')->with('error', ' Data Gagal Di Hapus');
+        }
         $data->delete();
-        return redirect()->route('anggota')->with('success', ' Data Berhasil Di Delete');
+        return redirect()->route('anggota')->with('error', ' Data Berhasil Di Hapus');
     }
 }
